issue #311: fix UI 404 when browsing table detail
JAX-RS path-matching gave priority to IndexesResource
(class path tablespaces/{tablespace}/tables/{table}) over
TablesResource's sub-path handler for requests like
GET /tablespaces/herd/tables/vector_bench, because the former
fully consumed all four path segments. Since IndexesResource
had no @GET at its class root, Jersey returned 404 for every
table-detail request, both for system tables and user tables.

Fix:
- Rename IndexesResource -> TableResource; it now owns the full
  single-table resource at tablespaces/{tablespace}/tables/{table}
- Add @GET at the TableResource class root that loads and returns
  table detail (columns, secondary indexes, foreign keys) using
  QueryService SQL queries; no delegation, no dead code
- Remove the now-unreachable get() method and its exclusive helpers
  from TablesResource; TablesResource is left with only list()
- Update ApplicationConfigV2 to register TableResource
- Rename IndexesResourceTest -> TableResourceTest; add four new
  direct-call tests for the table-detail endpoint
- Remove the four now-stale get* tests from TablesResourceTest
- Add HTTP smoke test in HealthResourceHttpSmokeTest that drives
  the full Jetty+Jersey routing for tablespaces/herd/tables/{name}
  to prevent regressions
