+++ herddb-ui/src/test/java/org/herddb/ui/api/v2/TablesResourceTest.java
@@ -30,12 +30,7 @@ import herddb.model.TableSpace;
 import herddb.model.TransactionContext;
 import java.util.Collections;
 import java.util.List;
-import javax.ws.rs.NotFoundException;
 import javax.ws.rs.WebApplicationException;
-import org.herddb.ui.dto.ColumnDTO;
-import org.herddb.ui.dto.ForeignKeyDTO;
-import org.herddb.ui.dto.IndexSummaryDTO;
-import org.herddb.ui.dto.TableDetailDTO;
 import org.herddb.ui.dto.TableSummaryDTO;
 import org.herddb.ui.internal.QueryService;
 import org.junit.Before;
@@ -43,10 +38,9 @@ import org.junit.Rule;
 import org.junit.Test;
 
 /**
- * Direct-call test of {@link TablesResource} against a real embedded
- * server. The {@code @Before} hook seeds a small schema with a parent
- * table, a child table that references it via FK, a secondary index, and
- * a few rows so the {@code systablestats} counters are non-zero.
+ * Direct-call tests for {@link TablesResource}: verifies the table-listing
+ * endpoint against a real embedded server. Individual table metadata (columns,
+ * indexes, foreign keys) is covered by {@link TableResourceTest}.
  */
 public class TablesResourceTest {
 
@@ -124,91 +118,13 @@ public class TablesResourceTest {
     }
 
     @Test
-    public void getReturnsColumnsIndexesAndForeignKeys() {
-        TablesResource resource = new TablesResource(
-                new QueryService(serverRule.getServer()));
-
-        TableDetailDTO detail = resource.get(TableSpace.DEFAULT, "child_t");
-
-        assertNotNull(detail);
-        assertNotNull(detail.getSummary());
-        assertEquals("child_t", detail.getSummary().getName().toLowerCase(java.util.Locale.ROOT));
-
-        // Columns: id, parent_id, amount in declaration order.
-        assertEquals(3, detail.getColumns().size());
-        assertEquals(
-                "id",
-                detail.getColumns().get(0).getName().toLowerCase(java.util.Locale.ROOT));
-        for (int i = 0; i < detail.getColumns().size(); i++) {
-            ColumnDTO column = detail.getColumns().get(i);
-            assertNotNull(column.getDataType());
-            assertNotNull(column.getTypeName());
-        }
-
-        // Index: child_t_amount_idx should exist (we only assert the
-        // user-defined one rather than enumerating implicit PK indexes).
-        boolean hasUserIndex = detail.getIndexes().stream()
-                .anyMatch(i -> "child_t_amount_idx".equalsIgnoreCase(i.getName()));
-        assertTrue(
-                "must include user-defined secondary index, got " + detail.getIndexes(),
-                hasUserIndex);
-        IndexSummaryDTO userIndex = detail.getIndexes().stream()
-                .filter(i -> "child_t_amount_idx".equalsIgnoreCase(i.getName()))
-                .findFirst()
-                .orElseThrow();
-        assertNotNull(userIndex.getColumns());
-        assertEquals(1, userIndex.getColumns().size());
-        assertEquals(
-                "amount",
-                userIndex.getColumns().get(0).toLowerCase(java.util.Locale.ROOT));
-
-        // Foreign key: child_t.parent_id -> parent_t.id.
-        assertFalse(
-                "child_t must have at least one FK, got " + detail.getForeignKeys(),
-                detail.getForeignKeys().isEmpty());
-        ForeignKeyDTO fk = detail.getForeignKeys().get(0);
-        assertEquals(
-                "parent_t",
-                fk.getParentTable().toLowerCase(java.util.Locale.ROOT));
-        assertEquals(List.of("parent_id"),
-                fk.getChildColumns().stream()
-                        .map(c -> c.toLowerCase(java.util.Locale.ROOT))
-                        .toList());
-        assertEquals(List.of("id"),
-                fk.getParentColumns().stream()
-                        .map(c -> c.toLowerCase(java.util.Locale.ROOT))
-                        .toList());
-    }
-
-    @Test
-    public void getReturns404ForUnknownTable() {
-        TablesResource resource = new TablesResource(
-                new QueryService(serverRule.getServer()));
-
-        assertThrows(
-                NotFoundException.class,
-                () -> resource.get(TableSpace.DEFAULT, "missing_t"));
-    }
-
-    @Test
-    public void getRejectsIllegalTablespace() {
+    public void listRejectsIllegalTablespace() {
         TablesResource resource = new TablesResource(
                 new QueryService(serverRule.getServer()));
 
         WebApplicationException ex = assertThrows(
                 WebApplicationException.class,
-                () -> resource.get("herd; DROP TABLE bad", "child_t"));
-        assertEquals(400, ex.getResponse().getStatus());
-    }
-
-    @Test
-    public void getRejectsIllegalTableName() {
-        TablesResource resource = new TablesResource(
-                new QueryService(serverRule.getServer()));
-
-        WebApplicationException ex = assertThrows(
-                WebApplicationException.class,
-                () -> resource.get(TableSpace.DEFAULT, "child_t; DROP"));
+                () -> resource.list("herd; DROP TABLE bad"));
         assertEquals(400, ex.getResponse().getStatus());
     }
 }
